moved calculate_desired_heading... test to the right test file
