+++ src/local_pathfinding/test/test_local_path.py
@@ -62,6 +62,76 @@ def create_test_local_path_for_in_collision_zone(
 
 
 @pytest.mark.parametrize(
+    "path, target_wp_index, boat_lat_lon, correct_heading, new_target_wp_index",
+    [
+        (
+            Path(
+                waypoints=[
+                    HelperLatLon(latitude=1.0, longitude=1.0),
+                    HelperLatLon(latitude=0.0, longitude=0.0),
+                ]
+            ),
+            1,
+            HelperLatLon(latitude=0.0, longitude=-0.1),
+            90.0,
+            1,
+        ),
+        (
+            Path(
+                waypoints=[
+                    HelperLatLon(latitude=1.0, longitude=1.0),
+                    HelperLatLon(latitude=0.0, longitude=0.0),
+                ]
+            ),
+            1,
+            HelperLatLon(latitude=0.1, longitude=0.0),
+            180.0,
+            1,
+        ),
+        (
+            Path(
+                waypoints=[
+                    HelperLatLon(latitude=1.0, longitude=1.0),
+                    HelperLatLon(latitude=0.0, longitude=0.0),
+                ]
+            ),
+            1,
+            HelperLatLon(latitude=0.1, longitude=0.1),
+            -135.0,
+            1,
+        ),
+        (
+            # Test: boat has reached waypoints[1], heading should be to waypoints[2].
+            Path(
+                waypoints=[
+                    HelperLatLon(latitude=0.0, longitude=0.2),
+                    HelperLatLon(latitude=0.0, longitude=0.1),
+                    HelperLatLon(latitude=0.0, longitude=0.0),
+                ]
+            ),
+            1,
+            HelperLatLon(latitude=0.0, longitude=0.09999),
+            -90.0,
+            2,
+        ),
+    ],
+)
+def test_calculate_desired_heading_and_waypoint_index(
+    path: Path,
+    target_wp_index: int,
+    boat_lat_lon: HelperLatLon,
+    correct_heading: float,
+    new_target_wp_index: int,
+):
+    calculated_answer = lp.LocalPath.calculate_desired_heading_and_wp_index(
+        path, target_wp_index, boat_lat_lon
+    )
+
+    assert calculated_answer[0] == pytest.approx(correct_heading, abs=3e-1)
+    assert calculated_answer[1] == new_target_wp_index
+
+
+@pytest.mark.parametrize(
     "target_local_wp_index, reference_latlon, path, obstacles, result",
     [
         (
